fix(apple): generate the SIWA nonce in Rust (iOS SwiftPM has no CryptoKit)
The iOS CI build failed: CryptoKit does not resolve in the plugin's
SwiftPM build context (type 'Insecure' has no member 'SHA256'). Move the
nonce generation and hashing into a target-independent Rust module
(nonce.rs) covered by the known-answer-vector tests; the Swift side now
receives both values, puts the hash on the authorization request and
echoes the raw nonce back. No behavior change on macOS.
